+++ ModBot.Domain/Interfaces/ModelsInterfaces/IPunishment.cs
@@ -6,7 +6,7 @@ namespace ModBot.Domain.Interfaces.ModelsInterfaces
 {
     public interface IPunishment
     {
-        ulong Id { get; }
+        int Id { get; }
         int StrikesAmount {get;}
         int TimeOutUntil { get;  }
     }
